+++ quiz-app/src/main/java/com/skillstorm/quizapp/repos/QuestionRepo.java
@@ -2,8 +2,8 @@ package com.skillstorm.quizapp.repos;
 
 import org.springframework.data.jpa.repository.*;
 
-import com.skillstorm.quizapp.models.Question;;
+import com.skillstorm.quizapp.models.Questions;
 
-public interface QuestionRepo extends JpaRepository<Question, Long> {
+public interface QuestionRepo extends JpaRepository<Questions, Long> {
 // data paging and sorting,
 }
