+++ quiz-app/src/main/resources/application.properties
@@ -1,12 +1,12 @@
 spring.application.name=quiz-app
 # Database URL
-spring.datasource.url=
+spring.datasource.url=jdbc:mysql:http://database-1-instance-1.chet1jfzht8q.us-east-1.rds.amazonaws.com
 
 # Database username
-spring.datasource.username=
+spring.datasource.username=admin
 
 # Database password
-spring.datasource.password=
+spring.datasource.password=rootroot
 
 # JDBC driver class name (optional, but recommended to specify)
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
